backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
pandas
numpy
scikit-learn
snowflake
snowflake-connector-python
snowflake-connector-python[pandas]
python-dotenv